Fix § mapping in normal mode
The intended behavior upon pressing § in normal mode is the insertion of
a blank line below the cursor.

Prior to this commit, the command mapping § to the escape key took
precedence over the intended mapping. This commit fixes the bug by
making the command unapplicable in normal mode (see :h map-modes).