Add missing backslash to call to sed
Prior to this commit, any attempt at sourcing the autoload file resulted
in the following error:

    Error detected while processing ~/.vim/plugged/iTunes-current-track/
    autoload/current_track.vim
    line 69:
    E126: Missing :endfunction

After replacing all the :endfun lines with :endfunction and encountering
the same error upon sourcing, careful reading of the code lead me to
discover that about 30 lines before I forgot to add a backslash. This
commit fixes that.

Note that the :CurrentTrack command still doesn't work. 🙁
But at least, I get no errors when sourcing the file. 🙂