Fix 'backspace' for versions of Vim before 8.2.590
The nostop feature was added to Vim on patch 590 of Vim 8.2, at this
commit: https://github.com/vim/vim/commit/aa0489e12d227d24752cf16e4e970.

Thus, when sourcing the vimrc on files that did not include that patch,
one would get the following error:
    E474: Invalid argument: backspace+=nostop

This commit sets the last item of 'backspace' to 'nostop' for versions
of Vim after 8.2.0590, and to 'start' for older versions of Vim.