Enable termguicolors
I am getting tired of having to set the option every time I launch Vim
in the terminal (the diffs of my colorscheme are impossible to read even
in 256-colors in the terminal), and I've only ever used terminals which
allow to display every color (I don't use macOS's default terminal app).

If this ever causes me any problem, it's easy enough to turn off the
option (se notgc), or to revert the commit.