Replace a continue by a return statement
The old design used the try_to_get_vim_plug() function to install 2
things; thus, I stored the list of things to install into a dictionary
and used a loop to install each file. The function only installs
vim-plug itself now, so I removed the loop.

However, while doing so, I forgot to turn a continue statement into a
return statement, causing Vim to raise an error since the continue
statement is no longer contained in a loop. This commit fixes that.