Patch colorschemes for quick-scope
Without these patches, it can be hard or impossible to tell apart the
characters highlighted by the quick-scope plugin, rendering it useless.
I didn't want to make pull requests to the author of the colorschemes,
so I prefered to add the necessary theming on my vimrc.

When possible, I tried to link the quick-scope highlight groups to
existing highlight groups.