Disable 'number'
Leaving the option on makes Vim take 1 more column for the line number,
with extra information that I rarely need.

If I need to know the number of the line the cursor is in for some
reason, I can always look on the bottom right of the window thanks to
Airline, regardless of the mode, or I can use CTRL-g in normal mode to
display informations about the file (including the line number of the
line at the cursor) on the status line.