Add vim-abolish
I've been hesitating to add this plugin to my vimrc; I didn't like the
coupling of features that are useful to me (i.e. the ones I mentioned in
the comment above the inclusion of the plugin) and those that I am
indifferent towards (e.g. :Abolish). But the usefulness of the former
features trumped the inconvenience brought by the inclusion of the
:Abolish command in the end.