Use gx instead of cx for vim-exchange
I'm discovering Emacs, and the Emacs port of vim-exchange (called
evil-exchange) uses gx instead of cx to exchange regions.  In order to
make switching between the two editors easier, I chose to remap the
bindings of the first plugin, as evil-exchange has a sound reason to use
gx instead of cx.

This change conflicted with my previous mapping for closing tabs (gx),
so I changed that mapping as well.