Add :Writeable command
This command sets the 'modifiable' option (to allow modifying the
contents of a buffer) and unsets the 'readonly' option (to allow saving
those changes to the file associated to that buffer).  This command is
mostly useful for modifying a Vim help file, something that I do often
since I use Vim's help system as a place to take notes.